# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QK 教培 CRM 管理系统 API
  version: 1.1.12
  summary: 线索、商机、客户全流程管理接口
  description: |
    面向教培行业销售团队的后台管理接口，覆盖线索录入、分配、跟进、转商机、再分配、
    再跟进、转客户的全流程，以及部门、角色、用户、课程、活动等基础数据维护。

    ## 通用约定

    ### 统一响应结构
    所有接口返回 `Result` 对象：`code` 为 `1` 表示成功、`0` 表示失败。
    **业务失败同样返回 HTTP 200**，由调用方依据 `code` 判断；`msg` 为可直接展示给用户的中文提示。
    无业务数据时 `data` 为 `null`，且该字段会**整个省略**（服务端 `Result` 标注 `@JsonInclude(NON_NULL)`），
    判断成功请用 `code == 1`，不要依赖 `data` 字段是否存在。

    ### 认证
    除 `POST /login` 外，所有接口都要求在请求头 `token` 中携带登录接口签发的 JWT 令牌。
    令牌缺失、被篡改、已过期，或令牌对应的账号已被停用 / 删除时，接口直接返回 **HTTP 401 且响应体为空**，
    不返回 `Result` 结构。
    令牌默认有效期 24 小时。

    ### 授权（权限点模型）
    已登录不等于有权限。每个接口在代码里声明自己需要的**权限点**（如 `clue:track`、`user:delete`，
    完整目录见 `GET /permissions`），账号能不能调用取决于「它的角色被授予了哪些权限点」。
    角色是数据，可以自由增删改，**改角色名/标识不影响任何权限**；授权关系存在 `role_permission` 表，
    用 `GET /roles/{id}/permissions` 查看、`PUT /roles/{id}/permissions` 覆盖式配置（改完立即生效）。
    未声明权限的接口对所有人拒绝（默认拒绝），唯一例外是 `POST /login`。
    内置超级管理员角色（`role.is_super = 1`）天然拥有全部权限，不可删除、也不支持单独调整权限。
    登录响应里的 `permissions` 即当前账号的权限码列表，前端用它控制菜单与按钮（不要用 `roleLabel`）。

    ### 分页
    分页接口的查询参数为 `page`（页码，从 1 开始，默认 1）与 `pageSize`（每页条数，默认 10，**单页最大 200**），
    响应数据统一为 `{ "total": 总记录数, "rows": [...] }`。
    `page` / `pageSize` 必须为正整数：传空值、0、负数或超过 200 时返回 `code = 0`，`msg` 为对应的字段提示。

    ### 错误语义（HTTP 状态码）
    | 状态码 | 触发场景 | 响应体 |
    | --- | --- | --- |
    | `200` | 成功；以及**业务失败**（数据不存在、唯一键冲突、文件类型不合法、状态不允许等） | `Result`，失败时 `code = 0`，按 `msg` 提示 |
...

...

...

...

...

...

...

...
servers:
...
security:
  - token: []
...
paths:
...
  /users:
    get:
      tags: [用户管理]
      summary: 条件分页查询用户列表
      operationId: listUsers
      parameters:
        - name: name
          in: query
          description: 姓名，模糊匹配
          schema: { type: string }
        - name: status
          in: query
          description: 状态，1:正常, 0:停用
          schema: { type: integer, format: int32, enum: [0, 1] }
        - name: phone
          in: query
          description: 手机号，模糊匹配
          schema: { type: string }
        - name: deptId
          in: query
          description: 部门ID，精确匹配
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，按最后修改时间倒序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUserPage'
    post:
      tags: [用户管理]
      summary: 新增用户
      description: |
        请求体不包含密码时，使用默认密码 `123`；包含密码时按 `md5(用户名 + 明文密码)` 存储。
        该接口会写入一条操作日志。
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBody'
            example:
              username: lisi
              name: 李四
              phone: '13800138002'
              email: lisi@example.com
              gender: 1
              status: 1
              deptId: 4
              roleId: 2
              remark: 内容编辑
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [用户管理]
      summary: 修改用户信息
      description: |
        请求体中的 `password` 会被服务端忽略并置空，密码修改需走单独的重置流程。
        该接口会写入一条操作日志。
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /users/list:
    get:
      tags: [用户管理]
      summary: 查询所有用户（下拉框）
      operationId: listAllUsers
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 全量用户列表，按ID升序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUserList'
  /users/role/{roleLabel}:
    get:
      tags: [用户管理]
      summary: 根据角色标识查询用户
      operationId: listUsersByRole
      parameters:
        - name: roleLabel
          in: path
          required: true
          description: 角色标识，如 `clue_operator`、`business_operator`
          schema: { type: string }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 用户列表，**只返回正常状态（`status = 1`）的用户**（分配线索/商机的人员下拉数据源，停用账号不可被分配）；角色标识不存在时返回空数组
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUserList'
  /users/dept/{deptId}:
    get:
      tags: [用户管理]
      summary: 根据部门ID查询用户
      operationId: listUsersByDept
      parameters:
        - name: deptId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 用户列表，部门下无用户时返回空数组
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUserList'
  /users/{ids}:
    get:
      tags: [用户管理]
      summary: 根据ID查询用户（回显）
      description: 路径参数为单个用户ID。
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserIds'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 用户信息；不存在时 `code = 0`、`msg = 用户不存在`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUser'
    delete:
      tags: [用户管理]
      summary: 批量删除用户
      description: 支持单个与批量删除，路径参数可传多个ID。该接口会写入一条操作日志。
      operationId: deleteUsers
      parameters:
        - $ref: '#/components/parameters/UserIds'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /depts:
    get:
      tags: [部门管理]
      summary: 条件分页查询部门
      operationId: listDepts
      parameters:
        - name: name
          in: query
          description: 部门名称，模糊匹配
          schema: { type: string }
        - name: status
          in: query
          description: 状态，1:正常, 0:停用
          schema: { type: integer, format: int32, enum: [0, 1] }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，按最后修改时间倒序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDeptPage'
    post:
      tags: [部门管理]
      summary: 新增部门
      description: 部门名称唯一，重复时返回 `msg = 部门名称已存在`。该接口会写入一条操作日志。
      operationId: addDept
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeptBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [部门管理]
      summary: 修改部门
      description: 该接口会写入一条操作日志。
      operationId: updateDept
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeptBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /depts/list:
    get:
      tags: [部门管理]
      summary: 查询所有正常状态的部门（下拉框）
      operationId: listAllDepts
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 仅返回 `status = 1` 的部门，按ID升序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDeptList'
  /depts/{id}:
    get:
      tags: [部门管理]
      summary: 根据ID查询部门
      operationId: getDeptById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 部门信息；不存在时 `code = 0`、`msg = 部门不存在`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDept'
    delete:
      tags: [部门管理]
      summary: 删除部门
      description: 该接口会写入一条操作日志。
      operationId: deleteDept
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /roles:
    get:
      tags: [角色管理]
      summary: 条件分页查询角色
      operationId: listRoles
      parameters:
        - name: name
          in: query
          description: 角色名称，模糊匹配
          schema: { type: string }
        - name: label
          in: query
          description: 角色标识，模糊匹配
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，按最后修改时间倒序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRolePage'
    post:
      tags: [角色管理]
      summary: 新增角色
      description: 角色标识全局唯一，重复时返回 `msg = 角色标识已存在`。该接口会写入一条操作日志。
      operationId: addRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [角色管理]
      summary: 修改角色
      description: 该接口会写入一条操作日志。
      operationId: updateRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /roles/list:
    get:
      tags: [角色管理]
      summary: 查询所有角色（下拉框）
      operationId: listAllRoles
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 全量角色列表，按ID升序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRoleList'
  /roles/{id}:
    get:
      tags: [角色管理]
      summary: 根据ID查询角色
      operationId: getRoleById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 角色信息；不存在时 `code = 0`、`msg = 角色不存在`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRole'
    delete:
      tags: [角色管理]
      summary: 删除角色
      description: 该接口会写入一条操作日志。
      operationId: deleteRole
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /roles/{id}/permissions:
    get:
      tags: [角色管理]
      summary: 查询角色已授予的权限
      description: |
        返回该角色当前拥有的权限点，按权限目录顺序排列。
        超级管理员角色返回全量目录（它天然拥有全部权限，界面上应展示为"全部勾选且不可编辑"）。
        页面渲染做法：先取 `GET /permissions` 拿到全量目录，再用本接口的返回回显勾选状态。
      operationId: findRolePermissions
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '200':
          description: 角色权限列表；角色不存在时 `code = 0`、`msg = 角色不存在`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPermissionList'
    put:
      tags: [角色管理]
      summary: 配置角色权限（覆盖式）
      description: |
        **覆盖式**更新：以本次提交的 `permissions` 为准，传空数组表示收回该角色的全部权限。
        权限码必须存在于权限目录（`GET /permissions`），出现未知权限码时返回 `code = 0`、
        `msg = 存在无效的权限码：<列表>`。
        超级管理员角色不支持单独调整（返回 `code = 0`、`msg = 超级管理员角色天然拥有全部权限，不支持单独调整`）。
        配置完成后立即对所有已登录的该角色账号生效（权限每次请求都会重新读取）。
        该接口会写入一条操作日志。
      operationId: updateRolePermissions
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermissionBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '200':
          $ref: '#/components/responses/Ok'
...
components:
  securitySchemes:
    token:
      type: apiKey
      in: header
      name: token
      description: |
        登录接口返回的 JWT 令牌。缺失、被篡改或已过期时接口返回 HTTP 401 且响应体为空。
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: 主键ID，64 位整型（对应数据库 bigint unsigned）
      schema: { type: integer, format: int64 }
    UserIds:
      name: ids
      in: path
      required: true
      description: |
        用户ID。查询回显时传单个ID，如 `5`；批量删除时可传多个ID，用英文逗号分隔，如 `1,2,3`。
      schema: { type: string }
      example: '1,2,3'
    Page:
      name: page
      in: query
      description: 页码，从 1 开始
      schema: { type: integer, format: int32, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: 每页记录数
      schema: { type: integer, format: int32, minimum: 1, default: 10 }
  responses:
    Ok:
      description: 操作成功，`code = 1`、`data` 为 `null`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
    Unauthorized:
      description: 令牌缺失、被篡改或已过期。响应体为空，不返回 `Result` 结构。
...
    Forbidden:
      description: |
        已登录但缺少该接口要求的权限点，返回 HTTP 403。
        响应体是 `Result`：`code = 0`，`msg` 为「无权访问该接口，请联系管理员分配权限」。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
  schemas:
...
    Result:
      type: object
      description: 统一响应结构。业务失败同样返回 HTTP 200，由 `code` 区分。
      properties:
        code:
          type: integer
          description: 1:成功, 0:失败
          enum: [1, 0]
        msg:
          type: string
          description: 提示信息，成功时固定为 `success`
        data:
          description: 业务数据，无数据时为 `null`
      required: [code, msg]
    DateTime:
      type: string
      description: 时间，输出格式 `yyyy-MM-dd HH:mm:ss`
      example: '2025-06-01 00:00:00'

    # ---------- 枚举 ----------
...
    Gender:
      type: integer
      format: int32
      description: 性别，1:男, 2:女
      enum: [1, 2]
...
    EnableStatus:
      type: integer
      format: int32
      description: 状态，1:正常, 0:停用
      enum: [0, 1]

    # ---------- 认证 ----------
...
    UserVO:
      type: object
      description: 用户视图对象，不含密码字段
      properties:
        id: { type: integer, format: int64, description: 用户ID }
        username: { type: string, description: 用户名 }
        name: { type: string, description: 姓名 }
        phone: { type: string, description: 手机号 }
        email: { type: string, description: 邮箱 }
        gender: { $ref: '#/components/schemas/Gender' }
        status: { $ref: '#/components/schemas/EnableStatus' }
        deptId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 部门ID
        roleId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 角色ID
        image:
          type: [string, 'null']
          description: 头像访问路径
        remark:
          type: [string, 'null']
          description: 备注
        createTime: { $ref: '#/components/schemas/DateTime' }
        updateTime: { $ref: '#/components/schemas/DateTime' }
        deptName:
          type: [string, 'null']
          description: 部门名称，由部门表关联查询得到
        roleName:
          type: [string, 'null']
          description: 角色名称，由角色表关联查询得到
    UserBody:
      type: object
      description: 新增或修改用户的请求体
      properties:
        id:
          type: integer
          format: int64
          description: 用户ID，修改时必填
        username: { type: string, description: 用户名，唯一 }
        password:
          type: string
          description: |
            明文密码。新增时留空则使用默认密码 `123`，否则按 `md5(用户名 + 明文密码)` 存储；
            修改时该字段被服务端忽略。
        name: { type: string, description: 姓名 }
        phone: { type: string, description: 手机号，唯一 }
        email: { type: string, description: 邮箱，唯一 }
        gender: { $ref: '#/components/schemas/Gender' }
        status: { $ref: '#/components/schemas/EnableStatus' }
        deptId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 部门ID
        roleId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 角色ID
        image:
          type: [string, 'null']
          description: 头像访问路径
        remark:
          type: [string, 'null']
          description: 备注
      required: [username, name, phone, email, gender, status]

    # ---------- 部门 ----------
    DeptVO:
      type: object
      properties:
        id: { type: integer, format: int64, description: 部门ID }
        name: { type: string, description: 部门名称，唯一 }
        status: { $ref: '#/components/schemas/EnableStatus' }
        createTime: { $ref: '#/components/schemas/DateTime' }
        updateTime: { $ref: '#/components/schemas/DateTime' }
    DeptBody:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 部门ID，修改时必填
        name: { type: string, description: 部门名称，唯一 }
        status: { $ref: '#/components/schemas/EnableStatus' }
      required: [name, status]

    # ---------- 角色 ----------
    RoleVO:
      type: object
      properties:
        id: { type: integer, format: int64, description: 角色ID }
        name: { type: string, description: 角色名称 }
        label: { type: string, description: 角色标识，全局唯一；仅用于展示与业务筛选，不参与接口授权 }
        remark:
          type: [string, 'null']
          description: 备注说明
        superRole:
          type: [boolean, 'null']
          description: 是否超级管理员角色；true 表示天然拥有全部权限，不可删除、不支持单独调整权限
        createTime: { $ref: '#/components/schemas/DateTime' }
        updateTime: { $ref: '#/components/schemas/DateTime' }
    RoleBody:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 角色ID，修改时必填
        name: { type: string, description: 角色名称 }
        label: { type: string, description: 角色标识，全局唯一；改名不影响已授予的权限 }
        remark:
          type: [string, 'null']
          description: 备注说明
      required: [name, label]
    PermissionVO:
      type: object
      description: 权限点
      properties:
        code: { type: string, description: 权限码，如 `clue:track` }
        description: { type: string, description: 权限名称，如「跟进线索」 }
    RolePermissionBody:
      type: object
      description: 覆盖式配置角色权限的入参
      properties:
        permissions:
          type: array
          description: 权限码列表；传空数组表示收回该角色的全部权限
          items: { type: string, example: 'clue:track' }
      required: [permissions]

    # ---------- 课程 ----------
...
    UserPage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/UserVO' }
      required: [total, rows]
    DeptPage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/DeptVO' }
      required: [total, rows]
    RolePage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/RoleVO' }
      required: [total, rows]
...
    ResultUserPage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/UserPage' }
    ResultUserList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/UserVO' }
    ResultUser:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/UserVO'
                - type: 'null'
    ResultDeptPage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/DeptPage' }
    ResultDeptList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/DeptVO' }
    ResultDept:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/DeptVO'
                - type: 'null'
    ResultRolePage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/RolePage' }
    ResultRoleList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/RoleVO' }
    ResultRole:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/RoleVO'
                - type: 'null'
    ResultPermissionList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/PermissionVO' }
...